re #175 fixing converter maps

+++ src/FlatBufferCreator.cpp
@@ -3,7 +3,7 @@
 namespace FlatBufs {
 
 void FlatBufferCreator::config(
-    std::map<std::string, int64_t> const &config_ints,
+    std::map<std::string, int> const &config_ints,
     std::map<std::string, std::string> const &config_strings) {
   UNUSED_ARG(config_ints);
   UNUSED_ARG(config_strings);
