+++ src/FlatBufferCreator.h
@@ -15,7 +15,7 @@ public:
   virtual ~FlatBufferCreator() = default;
   virtual std::unique_ptr<FlatbufferMessage>
   create(EpicsPVUpdate const &up) = 0;
-  virtual void config(std::map<std::string, int64_t> const &config_ints,
+  virtual void config(std::map<std::string, int> const &config_ints,
                       std::map<std::string, std::string> const &config_strings);
   virtual std::map<std::string, double> getStats();
 };
